Allow this to work inside tab view controllers
How? If the view will appear in a view controller that will use a tab bar, enable the "inTabViewController" inspector
attribute to prevent issues

Why? A view anchored to the top of a tab bar has its position compromised because the tab view animates downwards when the keyboard appears. This position, where the top of the tab bar is at the bottom of the visible screen results in a view that is slightly hidden by the tab bar. 

More briefly, when the tab bar animates away, the topOffset is the distance to the bottom of the screen and not to the top of tab bar.